Add GH workflow to build container images
We have a Dockerfile in the repository, so it would be nice to make container images of the bot automatically available when changes are merged.

Since the bot follows no explicit versioning scheme, I configured the built images to be tagged only with the 'edge' tag (to identify the latest commit) and with the short hash of the built commit.